Глобальное обновление: -Была добавлена система распределения работников по правам. Есть три основные категории: -Администраторы - работники со всемии правами редактирования/изменения/удаления других пользователей. -Модератор - работники с правами редактирования/изменения других пользователей. -Пользователь - работники с правами просмотра таблиц производства. (В будущем будет добавлена функция формирования отчёта (например заказа или найма сотрудников за данный период) -Переработан GUI и разделение его для работников с разными правами -Соответсвенно в таблице employees было добавлено поле "Права" -К данному полю был создан фильтр в контроллере EmployeeController.java Изменения система авторизации: -Добавлен вывод на экран id (паспорт) сессии -Происходит проверка работника на наличие тех или иных прав, вследствие чего открывается то или иное окно с соответсвующим для прав функционалом. Недоделано: -Для модераторов не были созданы и протестированы GUI -Недобавлен функционал редактирования/изменения существующих пользователей -Неубран весь шаблонный код из уже существующих классов

+++ src/main/java/org/example/demotest/repository/EmployeeRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 @Repository
 public interface EmployeeRepository extends JpaRepository<Employee, Long> {
     Employee findEmployeeById(Long id);
-    Optional<Employee> findByPassportNumber(Integer passportNumber);
+    Optional<Employee> findByPassportNumber(Long passportNumber);
     Employee findByPassportNumberAndPassWord(Long passportNumber, String password);
 
     Optional<Employee> findByPhoneNumber(String phoneNumber);
